Remove Dictionary, Linux changes
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@7051 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ phreeqc/structures.cpp
@@ -1763,7 +1763,7 @@ cxxChemRxn2rxn(cxxChemRxn &cr)
  *
  *   Return: rxn_ptr_new,  pointer to new structure 
  */
-	for (int i = 0; i < cr.Get_tokens().size(); i++)
+	for (int i = 0; i < (int) cr.Get_tokens().size(); i++)
 	{
 		cr.Get_tokens()[i].s = s_store(cr.Get_tokens()[i].s->name, cr.Get_tokens()[i].s->z, FALSE);
 		if (cr.Get_tokens()[i].name != NULL)
